requirements.txt
alembic==1.16.2
blinker==1.9.0
bottle==0.13.4
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
proxy_tools==0.1.0
pycairo==1.28.0
PyGObject==3.52.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
python-dotenv==1.1.1
pywebview==5.4
QtPy==2.4.3
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
